Feat/add tests

+++ ImageRepo.java
@@ -12,32 +12,42 @@ public class ImageRepo {
     static String display = "display";
 
     public static void main(String[] args) {
+
+        if(args.length == 0){
+            System.err.println("Please input a valid command, use -help or -h to find out about the commands");
+            return;
+        }
+
         String command = args[0];
-        /*
-        if(command.equals("-encrypt") || command.equals("-en")){
-            encryptFile(args[1]);
-        }
-        else if(command.equals("-decrypt") || command.equals("-de")){
-            decryptFile(args[1]);
-        }
-        */
         if(command.equals("-add") || command.equals("-a")){
+            if(args.length < 2){
+              System.err.println("Please input file path, use -help flag for command help");
+              return;
+            }
             addFile(args[1]);
         }
         else if(command.equals("-list") || command.equals("-l")){
             listFile();
         }
         else if(command.equals("-display") || command.equals("-dis")){
+            if(args.length < 2){
+              System.err.println("Please input file name, use -help flag for command help");
+              return;
+            }
             displayFile(args[1]);
         }
         else if(command.equals("-delete") || command.equals("-del")){
+            if(args.length < 2){
+              System.err.println("Please input file name, use -help flag for command help");
+              return;
+            }
             deleteFile(args[1]);
         }
         else if(command.equals("-help") || command.equals("-h")){
             displayHelp();
         }
         else{
-            System.out.println("Please input a valid command, use -help or -h to find out about the commands");
+            System.err.println("Please input a valid command, use -help or -h to find out about the commands");
         }
 
     }
@@ -45,7 +55,7 @@ public class ImageRepo {
     private static void addFile(String filename) {
         File path = new File(database);
         if(!Files.exists(Paths.get(database))){
-            return;
+            path.mkdir();
         }
         encryptFile(filename);
     }
